pyproject.toml
[project]
name = "dog-breed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "kaggle>=1.8.3",
    "matplotlib>=3.10.8",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pytorch-lightning>=2.6.0",
    "scikit-learn>=1.8.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
